[REFACTOR] #2 pakcage 변경 (ui, data 레이어 분리)
